Use indicator title as primary key for raster DB
Earlier I was using the prefix from a raster's filename as its primary
key for the raster DB. Also, I was saving the raster's metadata under a
key which was taken from user-supplied config.

Some errors were occurring because of the above choices. That's why, now
I'm using the title of the indicator in both places.

Also, I removed some unused code.

Signed-off-by: Abhineet Tamrakar <abhineet97@gmail.com>